requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-livereload-server==0.5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mpmath==1.3.0
mysql-connector==2.2.9
numpy==2.1.2
pyasn1==0.6.1
pycryptodome==3.21.0
PyJWT==2.9.0
rsa==4.9
scipy==1.14.1
sqlparse==0.5.1
sympy==1.13.3
tornado==6.4.1
typing_extensions==4.12.2
